Move console stream to the separate file.
