+++ Library/stdio.h
@@ -5,24 +5,8 @@
 #define size_t uint32_t
 
 #include <stdint.h>
-#include "micros.h"
-
-typedef enum file_buffering_mode
-{
-    file_buffering_mode_none,
-    file_buffering_mode_line,
-    file_buffering_mode_full
-} file_buffering_mode;
-
-typedef struct FILE
-{
-    char* buffer;
-    uint32_t pos;
-    file_buffering_mode buffering_mode;
-    
-    void (*fetch)(struct FILE* file);
-    void (*flush)(struct FILE* file);
-} FILE;
+#include "Streams/file.h"
+#include "Streams/console_stream.h"
 
 FILE *stdin;
 FILE *stdout;
@@ -41,8 +25,5 @@ int ungetc(int character, FILE* stream);
 size_t fwrite(const void* ptr, size_t size, size_t count, FILE* stream);
 
 FILE* streams_create_stream();
-FILE* streams_set_stream_as_console(FILE* stream);
-void streams_console_fetch(FILE* stream);
-void streams_console_flush(FILE* stream);
 
 #endif
